# Remove custom Utf8 encoder implementation